Each file is represented by an object still.
This time, it's a little smarter. Files are opened when the open(...) setter is called
with a value (currently, only 'ro' and 'rw' are valid). open() returns the current
permission type (undef if file is closed). open(undef) is the same as close(); the file
is closed. FileHandles may only be read/written if they are open(...)'d for that (this
may change in the future). Non-preexisting files (and their paths) are created only when
file is open('wo'). Hmmm... Ah. And I'm using 5.10's defined or (//=).

TODO: close the filehandle first before reopening it with different permissions.
TODO: do not re-open a file if the permissions are the same.